ProductCRUD/requirements.txt
django-cors-headers==3.13.0
Django==4.1.4
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
psycopg2-binary==2.9.5